Fix bf16 visualization crash: cast sampler output to float32 before NumPy
The W&B per-epoch visualization called .cpu().numpy() on the sampler output,
which is bf16 under autocast; NumPy has no bfloat16 dtype, raising
'Got unsupported ScalarType BFloat16'. Cast to float32 first (no-op for fp32).
The generation path (generator.py) already does this, so release output was
unaffected.

+++ wavedit/evaluation/visualization.py
@@ -52,7 +52,8 @@ def _dwt_roundtrip(image: torch.Tensor) -> torch.Tensor:
 
 
 def _to_unit_interval(volume: torch.Tensor) -> np.ndarray:
-    return np.clip((volume.cpu().numpy() + 1.0) / 2.0, 0.0, 1.0)
+    # .float() so bf16/fp16 sampler output converts to NumPy (no bfloat16 dtype there).
+    return np.clip((volume.float().cpu().numpy() + 1.0) / 2.0, 0.0, 1.0)
 
 
 @torch.no_grad()
